+++ packages/bridge-controller/src/utils/assets.ts
@@ -1,3 +1,4 @@
+import { KnownCaipNamespace, parseCaipAssetType } from '@metamask/utils';
 import type { CaipAssetType } from '@metamask/utils';
 
 import type { ExchangeRate, GenericQuoteRequest } from '../types.js';
@@ -38,4 +39,20 @@ export const toExchangeRates = (
     return acc;
   }, {});
   return exchangeRates;
+};
+
+export const assetIdsMatch = (
+  assetId1?: CaipAssetType,
+  assetId2?: CaipAssetType,
+): boolean => {
+  if (!assetId2 || !assetId1) {
+    return false;
+  }
+
+  return (
+    assetId1 === assetId2 ||
+    (parseCaipAssetType(assetId1).chain.namespace ===
+      KnownCaipNamespace.Eip155 &&
+      assetId1.toLowerCase() === assetId2.toLowerCase())
+  );
 };